moved all discord-specific utilities into their own file
