+++ src/discord-cluster-manager/cogs/admin_cog.py
@@ -13,15 +13,13 @@ import yaml
 from consts import GitHubGPU, ModalGPU
 from discord import app_commands
 from discord.ext import commands, tasks
+from discord_utils import leaderboard_name_autocomplete, send_discord_message, with_error_handling
 from leaderboard_db import LeaderboardItem, SubmissionItem
 from task import LeaderboardTask, make_task
 from ui.misc import ConfirmationView, DeleteConfirmationModal, GPUSelectionView
 from utils import (
     KernelBotError,
-    leaderboard_name_autocomplete,
-    send_discord_message,
     setup_logging,
-    with_error_handling,
 )
 
 if TYPE_CHECKING:
