+++ src/discord-cluster-manager/cogs/submit_cog.py
@@ -11,6 +11,7 @@ import discord
 from consts import GPU, GPU_TO_SM, RankCriterion, SubmissionMode, get_gpu_by_name
 from discord import app_commands
 from discord.ext import commands
+from discord_utils import send_discord_message, with_error_handling
 from report import (
     MultiProgressReporter,
     RunProgressReporter,
@@ -21,9 +22,7 @@ from run_eval import FullResult
 from task import LeaderboardTask, build_task_config
 from utils import (
     KernelBotError,
-    send_discord_message,
     setup_logging,
-    with_error_handling,
 )
 
 logger = setup_logging()
